Simplify source extraction
